+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 CHANGELOG
 =========
 
-1.0.0
+1.0.2
 -----
-* First public release.
+* Update TON SDK library to 1.2.0
+* Added default value (`null`) for some methods
+* Add new example: send rubies with comment to surf
+* Refactoring for the method `AbiType::fromJson()`
 
 1.0.1
 -----
-* Fix Extraton\TonClient\Net::queryCollection() orderBy parameter
+* Fixed parameter name for the method `Net::queryCollection()` 
+
+1.0.0
+-----
+* First public release.